I finally got the doubles to be ints.

+++ src/main/java/StudentDesks.java
@@ -16,12 +16,12 @@ class StudentDesks {
          *  your code goes here
          */
          double desks1 = students1 / 2.0;
-         desks1 = Math.round(desks1);
+         desks1 = (int)Math.round(desks1);
          double desks2 = students2 / 2.0;
-         desks2 = Math.round(desks2);
+         desks2 = (int)Math.round(desks2);
          double desks3 = students3 / 2.0;
-         desks3 = Math.round(desks3);
-         double totalDesks = desks1 + desks2 + desks3;
+         desks3 = (int)Math.round(desks3);
+         int totalDesks = (int)desks1 + (int)desks2 + (int)desks3;
          System.out.println(totalDesks);
          //This can only work if I use doubles.
         // closing the scanner object
